Add logging for curl errors in http
This will allow us to identify which codes we
potentially need to add extra checks for or
handle in different ways, such as retyring
for a timeout!

Remove unused http->encodeData stuff

These are legacy params that were used in
and old old old version of the framework.

We dont use them so we may as well get rid
of the extra complexity

Turn logging off for tests

This saves us from spamming the log files
that are also used for scripts with random
logs from tests that we dont actually care
about!

General tidy up of Logger class

Adding comments and stuff

All aimed at making the code more usable